requirements.txt
Django>=5.0
djangorestframework>=3.15.0
psycopg2-binary>=2.9.9
celery>=5.3.0
redis>=5.0.0
requests>=2.31.0
django-filter>=24.1
drf-spectacular>=0.27.1